- forgot to change the constructor to take double as well :D

+++ Rove/src/com/fourkins/rove/sqlite/posts/Post.java
@@ -12,7 +12,7 @@ public class Post {
 
     }
 
-    public Post(long id, String name, long latitude, long longitude, String message) {
+    public Post(long id, String name, double latitude, double longitude, String message) {
         super();
         this.id = id;
         this.userName = name;
